+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Rename some core functions
 - Support for MySQL port setting
 - Allow exported environment values to override `_ss_environment.php` values
+- Add DB type, DB function map, and errors for non-supported database types
 
 
 ## [0.0.2-alpha]